:fire: Déplace le bloc Résultat de l'aide à la déclaration dans un fichier séparé
